Split CSubNet out into netaddress.cpp
