+++ src/test/netbase_tests.cpp
@@ -3,6 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "netaddress.h"
 #include "netbase.h"
 #include "test/test_bitcoin.h"
